Move helper fn for enterprise package importing rule
